:game_die: Added P8SCII Tab
Needs to be formatted still, among many other changes. This was a quick push and I spent way to much time messing with highlight.js than I should have. It's not important. Need to get this updated to 0.2.2 - because a lot of changes have come in the past few months.